Added pop feature to twitter menu and updated references
